refactor: replace DropdownMenu exposes
